Work in progress

+++ src/main/java/com/automotive/models/entity/AutoEntity.java
@@ -1,7 +1,5 @@
 package com.automotive.models.entity;
 
-import com.automotive.models.entity.UserEntity;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -18,7 +16,7 @@ public class AutoEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
     private String brand;
     private String model;
     private String type;
@@ -28,4 +26,5 @@ public class AutoEntity {
     @ManyToOne()
     @JoinColumn(name = "ownedBy", referencedColumnName = "id")
     private UserEntity ownedBy;
+
 }
