+++ src/main/java/com/automotive/repository/AutoRepository.java
@@ -11,6 +11,6 @@ public interface AutoRepository extends JpaRepository<AutoEntity, Integer> {
 
     @Transactional
     @Query(nativeQuery = true, value = "Select * from auto where owned_By = ?1")
-    List<AutoEntity> findByUser(Long userID);
+    List<AutoEntity> findByUser(Integer userID);
 
 }
